Jump to near-completion vertion of "pvmctl"
In this step, we don't see "kvmctl", all are changed to "pvmctl".
In this version we use telnet port for "VM"s as monitoring console.
Some(many) commands are added to "p(k)vmctl" like of:
	* manage "VM" snapshots
	* "shutdown" & "reset" commands
	* create, rename, clone, delete ... "VM"s

In this version we use from a directory hierarchical and try to
put all of files in appropriate places.
This directories place in "PDNSoftCo. Base Directory: /etc/PDNSoftCo.bdir"