Jump to version 2.0, we have new config file (Arrays for disks and eths)
In this version, in the config file we used arrays for definition of
disks and eths. by this there isn't any limit on the number of defined
disks.
Also in the start function (do_start) we have some huristics in choose
of indexes for different disk types.
New commands are: mkiso (create iso file from cd in the cdrom), rmiso
mkbase (put the specified image to the BIs directory, this image may be
used as a base image), rmbase
chconf (specially for graphical interface, ability to change VM settings)