Use of UNIX Sockets to monitor VMs
Before this, we were using telnet ports (as monitor port) for communication with
the VMs. It was a big security hole.
In this version we are using UNIX Sockets as monitor ports.
Also in this commit we have pvm.conf file that was ommited in the previous
commit. This file contains all of the public varriables and general perpose
functions in the pvm environment.