Satisfy network down at stopping VMs ...
We had a bug, on some cases and when we shutting down a vm, PVM network
going unreachable for about 20 seconds ...
this problem was because of linux bridge behavior. bridge in all time
gets the smallest mac of attached interfaces to him. so when a tap device
(virtual machine device) has smallest mac and going down, bridge should
select other mac for himself and netwok of bridge going down about
30 seconds. to satify this, we changed the tap creation behavior.
In this commit, pvm creates "tap" interfaces and prepend to them "FE"
string to be the biggest mac in the set of bridge interfaces.

In this version there isn't any need to ifup-scripts for net interfaces.
attaching tap devices to "virtual switches" is done by pvmctl.
Also name of bridges corresponding to the VLan interfaces changed from
"brv" to "vsw".
Also "vlan[x]" tag in config file will accept only the switch(bridge) name
or an specified "ifup-script".